DOCS: docs/self_hosting/installation/kubernetes.mdx: Fix helm install command to include namespace (#782)
In `docs/self_hosting/installation/kubernetes.mdx`

The docs mention replacing `<your-namespace>`, but that's not part of
the `helm install` command that's shown. Added the namespace argument
there and changed the variable to `<namespace>` which seems like more of
a style match.

Also fix missing `>` on the earlier mention of `<namespace>`.